# Perf: Small performance improvement for CalculateChecksumForPath