--- a/src/Files.Shared/Helpers/ChecksumHelpers.cs
+++ b/src/Files.Shared/Helpers/ChecksumHelpers.cs
@@ -9,9 +9,9 @@ namespace Files.Shared.Helpers
 		public static string CalculateChecksumForPath(string path)
 		{
 			var buffer = Encoding.UTF8.GetBytes(path);
-			var hash = MD5.HashData(buffer);
-
-			return BitConverter.ToString(hash).Replace("-", string.Empty);
+			Span<byte> hash = stackalloc byte[MD5.HashSizeInBytes];
+			MD5.HashData(buffer, hash);
+			return Convert.ToHexString(hash);
 		}
 	}
 }