@onein528 2023-03-15
@yaira2 To use stackalloc will be helpful if they are going to calculate considerably long binary data like file content. In this case, they calculate path.